Fixes currency handling for USD-style decimals and related create/settings issues.
- Fix currency input so values like `29.95` are no longer rewritten as `2.995` (was stripping decimals and formatting with Indonesian locale)

- Store money amounts as `decimal` instead of integers so fractional currencies work across products, sales, purchases, and finance

- Fix supplier/customer create crash on PostgreSQL when unique email rule ignored an empty `id`

- Add **Create Setting** and **Load Defaults** on Settings (table was empty with no way to add currency keys)



## Related



- Fixes https://github.com/fajarghifar/inventory-management-system/issues/33



## Test plan



- [ ] Settings → **Load Defaults** creates currency + store rows

- [ ] Set USD format (`$`, `.`, `,`, `2` fraction digits), refresh page

- [ ] Create/edit product with purchase price `29.95` — value stays `29.95`

- [ ] Create supplier with email — no bigint `id <> ''` error

- [ ] Create customer with email — same as above

- [ ] Run migration `2026_07_17_000001_change_money_columns_to_decimal`

+++ app/DTOs/SaleData.php
@@ -19,9 +19,9 @@ readonly class SaleData
         public ?int $customer_id = null,
         public SaleStatus $status = SaleStatus::COMPLETED,
         public ?string $notes = null,
-        public int $cash_received = 0,
-        public int $change = 0,
-        public int $global_discount = 0,
+        public float $cash_received = 0,
+        public float $change = 0,
+        public float $global_discount = 0,
     ) {}
 
     public static function fromArray(array $data): self
@@ -34,9 +34,9 @@ readonly class SaleData
             customer_id: isset($data['customer_id']) ? (int) $data['customer_id'] : null,
             status: isset($data['status']) ? SaleStatus::from($data['status']) : SaleStatus::COMPLETED,
             notes: $data['notes'] ?? null,
-            cash_received: (int) ($data['cash_received'] ?? 0),
-            change: (int) ($data['change'] ?? 0),
-            global_discount: (int) ($data['global_discount'] ?? 0),
+            cash_received: (float) ($data['cash_received'] ?? 0),
+            change: (float) ($data['change'] ?? 0),
+            global_discount: (float) ($data['global_discount'] ?? 0),
         );
     }
